requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
PyAudio==0.2.14
python-engineio==4.3.0
python-socketio==5.5.0
yarl==1.9.4
